openai_vision/requirements.txt
# OpenAI Vision API サンプルプログラム - 必要なパッケージ

# OpenAI API
openai>=1.3.0

# 設定ファイル読み込み
pyyaml>=6.0

# 環境変数読み込み
python-dotenv>=1.0.0

# HTTP通信
requests>=2.28.0

# 型ヒント
typing-extensions>=4.5.0

# 非同期処理
aiohttp>=3.8.0